Use HTSlib API to set the CRAM reference
cram_load_reference() is an internal HTSlib function, and
shouldn't be called directly.  Instead use hts_set_opt() to
tell the CRAM reader which reference it should use.

As the (HTSlib-internal) cram/cram.h file no longer needs to
be included in ViralConsensus code, the change to CXXFLAGS
in commit 0b99fca can also be reverted.